requirements.txt
google-adk==1.14.0

google-cloud-firestore>=2.16.0,<3.0.0
google-cloud-storage>=2.18.0,<3.0.0
google-cloud-secret-manager>=2.20.0,<3.0.0
google-cloud-logging>=3.11.0,<4.0.0

google-auth>=2.29.0,<3.0.0
google-auth-oauthlib>=1.2.0,<2.0.0
google-api-python-client>=2.127.0,<3.0.0

python-dotenv==1.0.1
